refactor(opencode): split Git effect adapter
