+++ packages/opencode/src/effect/runtime.ts
@@ -1,7 +1,7 @@
 import { Effect, Layer, ManagedRuntime } from "effect"
 import { AccountEffect } from "@/account/effect"
 import { AuthEffect } from "@/auth/effect"
-import { Git } from "@/git"
+import { GitEffect } from "@/git/effect"
 import { Instances } from "@/effect/instances"
 import type { InstanceServices } from "@/effect/instances"
 import { TruncateEffect } from "@/tool/truncate-effect"
@@ -10,7 +10,7 @@ import { Instance } from "@/project/instance"
 export const runtime = ManagedRuntime.make(
   Layer.mergeAll(
     AccountEffect.defaultLayer, //
-    Git.defaultLayer,
+    GitEffect.defaultLayer,
     TruncateEffect.defaultLayer,
     Instances.layer,
   ).pipe(Layer.provideMerge(AuthEffect.layer)),
